refactor(editor): move main toolbar to editor
