+++ packages/editor/src/plugins/serlo-template-plugins/article.tsx
@@ -1,3 +1,4 @@
+import { SaveButton } from '@editor/editor-ui/editor-toolbar/save-button'
 import {
   type EditorPlugin,
   type EditorPluginProps,
@@ -10,7 +11,6 @@ import { ContentLoaders } from '@serlo/frontend/src/serlo-editor-integration/com
 import { editorContent, entity, entityType } from './common/common'
 import { EntityTitleInput } from './common/entity-title-input'
 import { MetadataFieldsModal } from './common/metadata-fields-modal'
-import { ToolbarMain } from './toolbar-main/toolbar-main'
 
 export const articleTypeState = entityType(
   {
@@ -55,7 +55,7 @@ function ArticleTypeEditor(props: EditorPluginProps<ArticleTypePluginState>) {
 
       <section itemProp="articleBody">{content.render()}</section>
 
-      <ToolbarMain showSubscriptionOptions {...props.state} />
+      <SaveButton showSubscriptionOptions {...props.state} />
     </>
   )
 }
